refactor: removing encode from viem package, put everything inside EncodingUtils
