+++ packages/v0-core/src/vault/index.ts
@@ -1,3 +1,4 @@
 export * from "./Vault"
 export * from "./VaultUtils"
+export * from "./EncodingUtils"
 export * from "./SettleData"
